+++ lib/lightning_web/router.ex
@@ -226,7 +226,7 @@ defmodule LightningWeb.Router do
 
         live "/w", WorkflowLive.Index, :index
         live "/w/new", WorkflowLive.Edit, :new
-        live "/w/new/collaborate", WorkflowLive.CollaborateNew, :new
+        live "/w/new/collaborate", WorkflowLive.Collaborate, :new
         live "/w/:id", WorkflowLive.Edit, :edit
         live "/w/:id/collaborate", WorkflowLive.Collaborate, :edit